backend/requirements.txt
Flask==2.2.5
flask-cors==4.0.1
flask_sqlalchemy==3.0.5
python-dotenv==1.0.0
pymysql==1.1.1
boto3==1.34.34
pyodbc==5.0.1
pytest==7.4.3
apscheduler==3.10.4
black==22.12.0
flake8==5.0.4
requests==2.28.1
numpy
pandas
python-dateutil
openpyxl
twilio==7.16.0